move PatchOperation struct to util/types/patch.go
