+++ pkg/util/types/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["pvc.go"],
+    srcs = [
+        "patch.go",
+        "pvc.go",
+    ],
     importpath = "kubevirt.io/kubevirt/pkg/util/types",
     visibility = ["//visibility:public"],
     deps = [
