+++ pkg/util/types/patch.go
@@ -17,9 +17,9 @@
  *
  */
 
-package mutators
+package types
 
-type patchOperation struct {
+type PatchOperation struct {
 	Op    string      `json:"op"`
 	Path  string      `json:"path"`
 	Value interface{} `json:"value,omitempty"`
