+++ pkg/virt-api/webhooks/mutating-webhook/mutators/BUILD.bazel
@@ -6,13 +6,13 @@ go_library(
         "migration-create-mutator.go",
         "namespace-limits.go",
         "preset.go",
-        "utils.go",
         "vm-mutator.go",
         "vmi-mutator.go",
     ],
     importpath = "kubevirt.io/kubevirt/pkg/virt-api/webhooks/mutating-webhook/mutators",
     visibility = ["//visibility:public"],
     deps = [
+        "//pkg/util/types:go_default_library",
         "//pkg/util/webhooks:go_default_library",
         "//pkg/virt-api/webhooks:go_default_library",
         "//pkg/virt-config:go_default_library",
@@ -41,6 +41,7 @@ go_test(
     embed = [":go_default_library"],
     deps = [
         "//pkg/testutils:go_default_library",
+        "//pkg/util/types:go_default_library",
         "//pkg/virt-api/webhooks:go_default_library",
         "//pkg/virt-config:go_default_library",
         "//pkg/virt-handler/node-labeller/util:go_default_library",
