+++ pkg/virt-handler/node-labeller/node_labeller.go
@@ -36,6 +36,7 @@ import (
 	kubevirtv1 "kubevirt.io/client-go/api/v1"
 	"kubevirt.io/client-go/kubecli"
 	"kubevirt.io/client-go/log"
+	utiltype "kubevirt.io/kubevirt/pkg/util/types"
 	virtconfig "kubevirt.io/kubevirt/pkg/virt-config"
 	"kubevirt.io/kubevirt/pkg/virt-handler/node-labeller/api"
 	"kubevirt.io/kubevirt/pkg/virt-handler/node-labeller/util"
@@ -183,19 +184,13 @@ func (n *NodeLabeller) run() error {
 }
 
 func (n *NodeLabeller) patchNode(originalNode, node *v1.Node) error {
-	type payload struct {
-		Op    string            `json:"op"`
-		Path  string            `json:"path"`
-		Value map[string]string `json:"value"`
-	}
-
-	p := make([]payload, 0)
+	p := make([]utiltype.PatchOperation, 0)
 	if !reflect.DeepEqual(originalNode.Labels, node.Labels) {
-		p = append(p, payload{
+		p = append(p, utiltype.PatchOperation{
 			Op:    "test",
 			Path:  "/metadata/labels",
 			Value: originalNode.Labels,
-		}, payload{
+		}, utiltype.PatchOperation{
 			Op:    "replace",
 			Path:  "/metadata/labels",
 			Value: node.Labels,
@@ -203,11 +198,11 @@ func (n *NodeLabeller) patchNode(originalNode, node *v1.Node) error {
 	}
 
 	if !reflect.DeepEqual(originalNode.Annotations, node.Annotations) {
-		p = append(p, payload{
+		p = append(p, utiltype.PatchOperation{
 			Op:    "test",
 			Path:  "/metadata/annotations",
 			Value: originalNode.Annotations,
-		}, payload{
+		}, utiltype.PatchOperation{
 			Op:    "replace",
 			Path:  "/metadata/annotations",
 			Value: node.Annotations,
